Extract models from server.manage to features
